REFACT: Move WellNetwork implementation to separate file
